pyproject.toml
[tool.poetry]
name = "event-horizon"
version = "0.1.0"
description = ""
authors = ["Julian Stephens <julian@julianstephens.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = { extras = ["async"], version = "^3.0.3" }
flask-sqlalchemy = "^3.1.1"
flask-caching = "^2.1.0"
flask-login = "^0.6.3"
flask-admin = "^1.6.1"
email-validator = "^2.1.1"
itsdangerous = "^2.2.0"
flask-wtf = "^1.2.1"
flask-script = "^2.0.6"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
apiflask = "^2.1.1"
bcrypt = "^4.1.2"
fastapi = { extras = ["all"], version = "^0.110.1" }
flask-cors = "^4.0.0"
flask-session = {extras = ["sqlalchemy"], version = "^0.8.0"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pydocstyle = "^6.3.0"
mypy = "^1.9.0"
pre-commit = "^3.7.0"


[tool.poetry.group.tests.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
coverage = "^7.4.4"
requests = "^2.31.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.23.6"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["event_horizon"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
